+++ apps/riven/lib/riven-settings.schema.ts
@@ -4,24 +4,8 @@ import dedent from "dedent";
 import { readFileSync } from "node:fs";
 import z from "zod";
 
-import packageJson from "../package.json" with { type: "json" };
+import { CorePlugins } from "./schemas/core-plugins.schema.ts";
 import { LogLevel } from "./utilities/logger/log-levels.ts";
-
-import type { Replace } from "type-fest";
-
-type CorePluginName = Replace<
-  Extract<keyof typeof packageJson.dependencies, `@repo/plugin-${string}`>,
-  "@repo/plugin-",
-  ""
->;
-
-export const CorePlugins = z.enum(
-  Object.keys(packageJson.dependencies)
-    .filter((dependency) => dependency.startsWith("@repo/plugin-"))
-    .map((dependency) =>
-      dependency.replace("@repo/plugin-", ""),
-    ) as CorePluginName[],
-);
 
 export const RivenSettings = z.object({
   attemptUnknownDownloads: z
